refactor: move métodos do Model para Trait
